Add NodeSelector to CR
Added a NodeSelector field to CR. Regenerated the deep copy and
CRD files using both make manifests and make generate.

Added NodeSelector to the Pod Spec under the Deployment template
on newDeploymentForGoRemote.

Also changed the samples podtracer.yaml file to use the nodeSelector
field and land on the test node.